Add UsedDevicesTab
