+++ src/lib/y2partitioner/widgets/pages/md_raid.rb
@@ -5,11 +5,12 @@ require "y2partitioner/device_graphs"
 require "y2partitioner/widgets/md_description"
 require "y2partitioner/widgets/configurable_blk_devices_table"
 require "y2partitioner/widgets/edit_blk_device_button"
+require "y2partitioner/widgets/used_devices_tab"
 
 module Y2Partitioner
   module Widgets
     module Pages
-      # A Page for a md raid device: contains {MdTab} and {MdUsedDevicesTab}
+      # A Page for a md raid device: contains {MdTab} and {UsedDevicesTab}
       class MdRaid < CWM::Page
         # Constructor
         #
@@ -45,7 +46,7 @@ module Y2Partitioner
             ),
             Tabs.new(
               MdTab.new(@md),
-              MdUsedDevicesTab.new(@md, @pager)
+              UsedDevicesTab.new(@md.devices, @pager)
             )
           )
         end
@@ -78,45 +79,6 @@ module Y2Partitioner
               MdDescription.new(@md),
               Left(HBox(EditBlkDeviceButton.new(device: @md)))
             )
-        end
-      end
-
-      # A Tab for devices used by given md raid device
-      class MdUsedDevicesTab < CWM::Tab
-        # Constructor
-        #
-        # @param md [Y2Storage::Md]
-        # @param pager [CWM::TreePager]
-        def initialize(md, pager)
-          textdomain "storage"
-          @md = md
-          @pager = pager
-        end
-
-        # @macro seeAbstractWidget
-        def label
-          _("&Used Devices")
-        end
-
-        # @macro seeCustomWidget
-        def contents
-          @contents ||= VBox(table)
-        end
-
-      private
-
-        # Returns a table with all devices used by a MD raid
-        #
-        # @return [ConfigurableBlkDevicesTable]
-        def table
-          return @table unless @table.nil?
-          @table = ConfigurableBlkDevicesTable.new(devices, @pager)
-          @table.show_columns(:device, :size, :format, :encrypted, :type)
-          @table
-        end
-
-        def devices
-          @md.devices
         end
       end
     end
